\name{SAVI}
\alias{SAVI}

\title{
SAVI - Sheffield Accelerated Value of Information
}
\description{
This function launches a local version of SAVI, the Sheffield Accelerated Value of Information R Shiny App.
}

\usage{
SAVI()
}

\author{
Mark Strong, Penny Breeze, Chloe Thomas and Alan Brennan.

Email us at \email{savi@sheffield.ac.uk}.
}

\references{
Strong M, Oakley JE, Brennan A. Estimating multi-parameter partial Expected Value of Perfect Information from a probabilistic sensitivity analysis sample: a non-parametric regression approach. Medical Decision Making. 2014;34(3):311-26.

\url{http://mdm.sagepub.com/content/34/3/311}
}

\seealso{
An online version of SAVI is available at \url{http://savi.shef.ac.uk/SAVI/}.
}

\examples{
\dontrun{
SAVI()
}
}
